Created siteRequestUtils.js and moved the repeated codes in sites' route codes to it. Also, updated the site names specifically.
